added TypeBuilder and TypeModifier, updated reflection, started working on Entity saving/loading

+++ Engine/src/MathUtils.h
@@ -52,11 +52,11 @@ namespace Ghurund {
         return (data[(unsigned int)x] << 24) | (data[(unsigned int)y] << 16) | (data[(unsigned int)z] << 8) | data[(unsigned int)w];
     }
 
-	constexpr unsigned int operator"" _KB(unsigned long long val) {
+	constexpr unsigned long long operator"" _KB(unsigned long long val) {
 		return val * 1024;
 	}
 
-	constexpr unsigned int operator"" _MB(unsigned long long val) {
+	constexpr unsigned long long operator"" _MB(unsigned long long val) {
 		return val * 1024 * 1024;
 	}
 }
